Creation of singular obstacles for more interesting gameplay. Oxygen system re-worked to increase and decrease using deltatime. Also some cleaning up of code.

+++ UI/OxygenMeter.cs
@@ -7,12 +7,12 @@ public class OxygenMeter : MonoBehaviour
 {
     public Slider oxygenSlider;
     
-    public void SetMaxOxygen(int oxygen)
+    public void SetMaxOxygen(float oxygen)
     {
         oxygenSlider.maxValue = oxygen;
         oxygenSlider.value = oxygen;
     }
-    public void SetOxygen (int oxygen)
+    public void SetOxygen (float oxygen)
     {
         oxygenSlider.value = oxygen;
     }
